Fix indexing between GPU and host.
ssize_t has a different size (sometimes? only happened on libgomp)
when passed to the GPU, resulting in meaningless indexes.

+++ src/RaiderSTREAM/Impl/RS_OMP_TARGET/RS_OMP_TARGET.h
@@ -25,7 +25,7 @@
 class RS_OMP_TARGET : public RSBaseImpl {
 private:
   std::string kernelName;
-  ssize_t streamArraySize;
+  long streamArraySize;
   int numPEs;
   int numTeams;
   int threadsPerTeam;
